[backend] db relations & logger

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { GetTasksFilterDto } from './dto/get-tasks-filter.dto'
 import { TaskRepository } from './task.repository'
 import { InjectRepository } from '@nestjs/typeorm'
 import { Task } from './task.entity'
+import { User } from 'src/auth/user.entity'
 
 @Injectable()
 export class TasksService {
@@ -13,13 +14,15 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto)
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user)
   }
 
-  async getTaskById(id: number): Promise<Task> {
+  async getTaskById(id: number, user: User): Promise<Task> {
     try {
-      const task = await this.taskRepository.findOne(id)
+      const task = await this.taskRepository.findOne({
+        where: { id, userId: user.id },
+      })
       if (!task) {
         throw new NotFoundException(`Task with ID "${id} not found"`)
       }
@@ -29,26 +32,26 @@ export class TasksService {
     }
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto)
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user)
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id)
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user)
     task.status = status
     return await task.save()
   }
 
-  async deleteTask(id: number): Promise<void> {
-    try {
-      const result = await this.taskRepository.delete(id)
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id })
 
-      // decided to explicitly state that it's gonna be a number, not bool
-      if (result.affected === 0) {
-        throw new NotFoundException(`Task with ID "${id} not found"`)
-      }
-    } catch (error) {
-      throw error
+    // decided to explicitly state that it's gonna be a number, not bool
+    if (result.affected === 0) {
+      throw new NotFoundException(`Task with ID "${id} not found"`)
     }
   }
 }
